tests/sb3files/blocktest/output/requirements.txt
pygame-ce
js2py
numpy